initial commit
developed up to the point, where you can move the coordinate system,
where it has a dragable grid and where you can place points too.
Everything was done with html canvas and react. The storage of
points (and probably many other thinks in the future) is done with redux and redux react.